[product]三个Tab的基本展示

+++ src/com/ec/mamiya/data/ProductReview.java
@@ -2,19 +2,19 @@ package com.ec.mamiya.data;
 
 public class ProductReview {
 
-    private int mScore;
+    private float mScore;
     private String mAccount;
     private String mReview;
     private String mParams;
     
-    public ProductReview(int score, String account, String review, String params) {
+    public ProductReview(float score, String account, String review, String params) {
         mScore = score;
         mAccount = account;
         mReview = review;
         mParams = params;
     }
     
-    public int getScore() {
+    public float getScore() {
         return mScore;
     }
